Trath: Interactive Website Development
This is a .NET dynamic website (Data-Driven Web Application) developed on visual studio which I spent two weeks making in undergraduate year 2. Pretty many functions and details are designed and practiced in this project. I was content and happy to have opportunity to take this project independently although it looks immature at the moment.